- Compile under VC10
git-svn-id: svn+ssh://localhost/home/svn/musador/trunk@166 abf014a6-d86a-7740-8175-6a5112e866a6

+++ src/3rdParty/taglib/mpeg/mpegheader.cpp
@@ -157,7 +157,7 @@ void MPEG::Header::parse(const ByteVector &data)
     return;
   }
 
-  std::bitset<32> flags(data.toUInt());
+  std::bitset<32> flags(static_cast<unsigned long long>(data.toUInt()));
 
   // Check for the second byte's part of the MPEG synch
